Import of JONIONS/PANT-0.14 from CPAN.
gitpan-cpan-distribution: PANT
gitpan-cpan-version:      0.14
gitpan-cpan-path:         JONIONS/PANT/PANT-0.14.tar.gz
gitpan-cpan-author:       JONIONS
gitpan-cpan-maturity:     released

+++ t/ptest.t
@@ -1,4 +1,4 @@
-use Test::More tests => 4;
+use Test::More tests => 9;
 BEGIN { use_ok('PANT') };
 BEGIN { use_ok('PANT::Test') };
 
@@ -8,14 +8,19 @@ my @testarg = ("-output", $outfile);
 @delfiles = ();
 StartPant();
 push(@delfiles, $outfile);
-ok(RunTests(tests=>[qw(t/fake.t)], directory=>"."), "Run tests completes ok");
+ok(RunTests(tests=>[qw(t/fake/fake.t t/fake/fake2.t)], directory=>"."), "Run tests completes ok");
 
 EndPant();
 
 my $contents = FileLoad($outfile);
-like($contents, qr{\d+/\d+ subtests failed, \d+\.\d*% okay}i, "Test summary appears");
+like($contents, qr{Summary: Test Files 2}i, "Test summary appears");
+like($contents, qr{Summary: Total Tests 20}i, "Test summary appears");
+like($contents, qr{Took\s+\d+\s+wallclock}i, "Benchmark appears");
+like($contents, qr{td[^>]+id="fail"}i, "Failure tag detected");
+like($contents, qr{td[^>]+id="pass"}i, "Pass tag detected");
+like($contents, qr{Looks like you failed}i, "Stderr junk detected");
+unlink(@delfiles) unless $ENV{DEBUGTEST};
 
-unlink(@delfiles);
 sub FileLoad {
     my $fname = shift;
     local(*INPUT, $/);
